fix: build warning (#950)
* fix: replaced Sass division with math.div or calc

* fix: updated import path to avoid circular dependency

* fix: moved isTreeNode and getPosition function to avoid circular dependency

* fix: moved FilterState and flattenKeys to avoid circular dependency

* fix: moved getTimeProps and toArray to avoid circular dependency

* fix: moved core Tree logic to BaseTree to avoid circular dependency

* fix: passing cascadingMenu using render function to avoid circular depedency

* fix: replace math.div with calc for CSS value

* fix: removed redundant type file
