+++ src/components/Avatar/Avatar.tsx
@@ -10,7 +10,7 @@ import {
   StatusItemIconAlign,
   StatusItemsPosition,
   StatusItemsProps,
-} from './';
+} from './Avatar.types';
 import { Icon } from '../Icon';
 import { Popup } from '../Popup';
 import { Tooltip } from '../Tooltip';
@@ -317,7 +317,7 @@ const AvatarIcon: FC<AvatarIconProps> = React.forwardRef(
       onMouseEnter,
       onMouseLeave,
       style,
-      tabIndex
+      tabIndex,
     },
     ref: Ref<HTMLDivElement>
   ) => {
