+++ src/components/DateTimePicker/DatePicker/Styles/rtl.scss
@@ -31,7 +31,7 @@
 
     .picker-input {
       .picker-suffix {
-        margin-right: ($space-xs / 2);
+        margin-right: calc($space-xs / 2);
         margin-left: 0;
       }
