+++ src/components/RadioButton/RadioGroup.tsx
@@ -4,7 +4,8 @@ import React, { FC, Ref, useContext } from 'react';
 import DisabledContext, { Disabled } from '../ConfigProvider/DisabledContext';
 import { SizeContext, Size, OcThemeName } from '../ConfigProvider';
 import ThemeContext from '../ConfigProvider/ThemeContext';
-import { RadioButtonProps, RadioGroupProps, RadioButton } from './';
+import { RadioButtonProps, RadioGroupProps } from './Radio.types';
+import { RadioButton } from './RadioButton';
 import {
   LabelAlign,
   LabelPosition,
